fix amp error in plotting

+++ the_well/benchmark/metrics/plottable_data.py
@@ -90,7 +90,7 @@ def field_histograms(
 
 
 def build_1d_power_spectrum(x, spatial_dims):
-    x_fft = torch.fft.fftn(x, dim=spatial_dims, norm="ortho").abs().square()
+    x_fft = torch.fft.fftn(x.to(torch.float32), dim=spatial_dims, norm="ortho").abs().square()
     # Return the shifted sqrt power spectrum
     # First average over spatial dims, then take the last time step from the first batch element
     if len(spatial_dims) > 1:
